hiro marker with embedded qr code converted to .png file
